Arrow class was moved to image.arrow subpackage and renamed to ArrowDrawer. MapList class was moved to image.circle subpackage and renamed to CircleDrawer. MapRectangle class was moved to image.rectangle subpackage and renamed to RectangleDrawer. TextDrawer and its implementation was moved to image.text subpackage
